+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## Unreleased
-
 - Improve support for [Apollo Federation](https://www.apollographql.com/docs/apollo-server/federation/introduction/) by instrumenting [@apollo/gateway](https://www.npmjs.com/package/@apollo/gateway).
+- Downgrade log level for missing package.json in potential depedency directory from info to debug.
 
 ## 1.86.0
 - Add instrumentation for [pg-native](https://www.npmjs.com/package/pg-native) (`pg-native` is an alternative PostgreSQL driver package, in addition to the already supported [pg](https://www.npmjs.com/package/pg) package).